% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateemail.R
\name{validate_email}
\alias{validate_email}
\title{validate email}
\usage{
validate_email(x)
}
\arguments{
\item{x}{}

\item{address}{ADDRESS for validation}
}
\value{
TRUE if address is valid
}
\description{
validate email
}
\examples{
validate_email("marylin@sas.upenn.edu")
validate_email("adasdadasdfmai.com")
}
